Added forceCalculationOnCPU
Even when available, bugs in the OpenCL code can render the whole module useless if we cannot bypass OpenCL. The user can disable it:

1. with the environement variable PYTISSUE_FORCE_CPU = 1
2. with the function `forceCalculationOnCPU`

It is not perfect, but at least it works.

+++ pytissueoptics/rayscattering/opencl/__init__.py
@@ -1,7 +1,9 @@
 from pytissueoptics.rayscattering.opencl.config.CLConfig import warnings, CLConfig, OPENCL_AVAILABLE, WEIGHT_THRESHOLD
 from pytissueoptics.rayscattering.opencl.config.IPPTable import IPPTable
+import os
 
 OPENCL_OK = True
+PYTISSUE_FORCE_CPU = os.environ.get('PYTISSUE_FORCE_CPU', '0')
 
 if OPENCL_AVAILABLE:
     try:
@@ -12,9 +14,16 @@ if OPENCL_AVAILABLE:
 else:
     CONFIG = None
 
+def forceCalculationOnCPU():
+    os.environ["PYTISSUE_FORCE_CPU"] = "1"
+    print("You can define PYTISSUE_FORCE_CPU=1 in your profile to avoid this call.")
 
 def validateOpenCL() -> bool:
     notAvailableMessage = "Error: Hardware acceleration not available. Falling back to CPU. "
+    
+    if os.environ.get("PYTISSUE_FORCE_CPU", '0') != '0':
+        warnings.warn("User requested not using OpenCL with environment variable 'PYTISSUE_FORCE_CPU'=1.")
+        return False
     if not OPENCL_AVAILABLE:
         warnings.warn(notAvailableMessage + "Please install pyopencl.")
         return False
@@ -27,4 +36,4 @@ def validateOpenCL() -> bool:
 
 
 def hardwareAccelerationIsAvailable() -> bool:
-    return OPENCL_AVAILABLE and OPENCL_OK
+    return OPENCL_AVAILABLE and OPENCL_OK and not PYTISSUE_FORCE_CPU
